feature : 상품 관리 API 개발
- 상품 등록 API
- 상품 상세 조회 API
- 상품 검색 API

+++ src/main/java/com/dino/cqrs_challenge/global/response/PaginatedApiResponse.java
@@ -31,7 +31,7 @@ public class PaginatedApiResponse<T> {
     public static class Pagination {
 
         @Schema(description = "전체 항목 수", example = "100")
-        private int totalItems;
+        private long totalItems;
 
         @Schema(description = "전체 페이지 수", example = "10")
         private int totalPages;
